requirements.txt
google-genai
opencv-python-headless
insightface
onnxruntime
aiosqlite
Pillow
numpy
streamlit
python-dotenv
